refactor(ts): migrate ClearRefinements component
