+++ src/components/ClearRefinements/__tests__/ClearRefinements-test.tsx
@@ -3,6 +3,7 @@
 import { h } from 'preact';
 import ClearRefinements from '../ClearRefinements';
 import { mount } from 'enzyme';
+import { ReactElementLike } from 'prop-types';
 
 describe('ClearRefinements', () => {
   const defaultProps = {
@@ -22,14 +23,18 @@ describe('ClearRefinements', () => {
   };
 
   it('should render <ClearRefinements />', () => {
-    const wrapper = mount(<ClearRefinements {...defaultProps} />);
+    const wrapper = mount(
+      (<ClearRefinements {...defaultProps} />) as ReactElementLike
+    );
 
     expect(wrapper).toMatchSnapshot();
   });
 
   it('should render <ClearRefinements /> with a specific class when no refinements', () => {
     const wrapper = mount(
-      <ClearRefinements {...defaultProps} hasRefinements={false} />
+      (
+        <ClearRefinements {...defaultProps} hasRefinements={false} />
+      ) as ReactElementLike
     );
 
     expect(wrapper).toMatchSnapshot();
